requirements.txt
langchain
langchain-openai
langchain-mcp-adapters
openai
python-dotenv
pydantic
requests
asyncio
typing-extensions 